+++ lib/segment/src/index/field_index/numeric_index/index.rs
@@ -8,7 +8,7 @@ use common::types::PointOffsetType;
 use gridstore::Blob;
 use serde_json::Value;
 
-use super::read_ops::NumericIndexRead;
+use super::numeric_index_read::NumericIndexRead;
 use super::storage::NumericIndexInner;
 use super::{Encodable, NumericIndexGridstoreBuilder, NumericIndexMmapBuilder};
 use crate::common::operation_error::OperationResult;
