+++ lib/segment/src/index/field_index/numeric_index/mutable_numeric_index/read_only/read_ops.rs
@@ -6,7 +6,7 @@ use common::universal_io::UniversalRead;
 use gridstore::Blob;
 
 use super::super::super::Encodable;
-use super::super::super::read_ops::NumericIndexRead;
+use super::super::super::numeric_index_read::NumericIndexRead;
 use super::ReadOnlyAppendableNumericIndex;
 use crate::common::operation_error::OperationResult;
 use crate::index::field_index::histogram::Histogram;
