+++ lib/segment/src/index/field_index/numeric_index/numeric_field_index.rs
@@ -1,7 +1,7 @@
 use common::types::PointOffsetType;
 use itertools::Either;
 
-use super::read_ops::NumericIndexRead;
+use super::numeric_index_read::NumericIndexRead;
 use super::{NumericIndexInner, StreamRange};
 use crate::common::operation_error::OperationResult;
 use crate::data_types::order_by::OrderValue;
