+++ lib/segment/src/index/field_index/numeric_index/storage/read_ops.rs
@@ -12,7 +12,7 @@ use common::types::PointOffsetType;
 use gridstore::Blob;
 
 use super::super::Encodable;
-use super::super::read_ops::NumericIndexRead;
+use super::super::numeric_index_read::NumericIndexRead;
 use super::NumericIndexInner;
 use crate::common::operation_error::OperationResult;
 use crate::index::field_index::histogram::Histogram;
